casa3aaaaaaaaaaaaaaaaaaaa

+++ EjTema4Part2CarlosOrtega/src/ej1/IImpuesto.java
@@ -2,7 +2,7 @@ package ej1;
 
 public interface IImpuesto {
 
-	double calculoIva(double precio, int iva);
+	double calculoIva(double precio, double iva);
 	
 	double calculoIrpf (double sueldo);
