+++ EjTema4Part2CarlosOrtega/src/ej2/IEstadisticas.java
@@ -2,7 +2,7 @@ package ej2;
 
 public interface IEstadisticas {
 	
-	double calcularMinimo();
-	double calcularMaximo();
-	double calcularSumatorio();
+	int calcularMinimo();
+	int calcularMaximo();
+	int calcularSumatorio();
 }
